repaired broken db dir after refactor
