+++ spindle_db/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "spindle_db"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2021"
 license = "Apache-2.0 OR MIT"
 description = "A db backend for `spindle`: ergonomic GPGPU data parallel computing in Rust."
