+++ spindle_macros/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "spindle_macros"
-version = "0.1.8"
+version = "0.1.9"
 edition = "2021"
 license = "Apache-2.0 OR MIT"
 description = "A macro backend for `spindle`: ergonomic GPGPU data parallel computing in Rust."
@@ -14,6 +14,6 @@ proc-macro2 = "1.0.64"
 quote = "1.0.29"
 serde = { version = "1.0.176", features = ["derive"] }
 serde_json = "1.0.104"
-spindle_db = { path = "../spindle_db" } # todo! this doesn't feel correct
+spindle_db = { version = "0.1.2", path = "../spindle_db" } # todo! this doesn't feel correct
 syn = { version = "2.0.25", features = ["full", "parsing", "printing"] }
 uuid = { version = "1.4.1", features = ["fast-rng", "macro-diagnostics", "serde", "v4"] }
